refactor: rename broadcaster to sender since not all sendings are broadcasting
