+++ src/net/mod.rs
@@ -1,2 +1,2 @@
-pub mod broadcaster;
 pub mod listener;
+pub mod sender;
